Clean up old custom style implementation.
The newer implementation is much simpler, so only some of the code needs
to be ported over. This is a WIP. The old tests don't work yet.
